requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
beautifulsoup4
requests
